+++ packages/api/tests/endpoints/granules/test-bulk-delete.js
@@ -34,6 +34,7 @@ test.before(async () => {
   process.env.TOKEN_SECRET = randomString();
   process.env.AccessTokensTable = randomString();
   process.env.GranulesTable = randomString();
+  process.env.granule_sns_topic_arn = randomString();
   process.env.METRICS_ES_HOST = randomString();
   process.env.METRICS_ES_USER = randomString();
   process.env.METRICS_ES_PASS = randomString();
@@ -115,6 +116,7 @@ test.serial('POST /granules/bulkDelete starts an async-operation with the correc
       cmr_provider: process.env.cmr_provider,
       cmr_username: process.env.cmr_username,
       GranulesTable: process.env.GranulesTable,
+      granule_sns_topic_arn: process.env.granule_sns_topic_arn,
       launchpad_api: process.env.launchpad_api,
       launchpad_certificate: process.env.launchpad_certificate,
       launchpad_passphrase_secret_name: process.env.launchpad_passphrase_secret_name,
@@ -172,6 +174,7 @@ test.serial('POST /granules/bulkDelete starts an async-operation with the correc
       cmr_provider: process.env.cmr_provider,
       cmr_username: process.env.cmr_username,
       GranulesTable: process.env.GranulesTable,
+      granule_sns_topic_arn: process.env.granule_sns_topic_arn,
       launchpad_api: process.env.launchpad_api,
       launchpad_certificate: process.env.launchpad_certificate,
       launchpad_passphrase_secret_name: process.env.launchpad_passphrase_secret_name,