+++ packages/db/src/translate/granules.ts
@@ -23,14 +23,13 @@ import { translatePostgresFileToApiFile } from './file';
  *
  * @param {Object} params
  * @param {PostgresGranuleRecord} params.granulePgRecord - Granule from Postgres
- * @param {PostgresCollectionRecord} params.collectionPgRecord - Collection from Postgres
+ * @param {PostgresCollectionRecord} [params.collectionPgRecord] - Optional Collection from Postgres
  * @param {Knex | Knex.Transaction} params.knexOrTransaction
  *   Knex client for reading from RDS database
- * @param {Object} params.collectionPgModel - Instance of the collection database model
- * @param {Object} params.pdrPgModel - Instance of the pdr database model
- * @param {Object} params.providerPgModel - Instance of the provider database model
- * @param {Object} params.filePgModel - Instance of the file database model
- * @param {Object} params.executionPgModel - Instance of the execution database model
+ * @param {Object} [params.collectionPgModel] - Instance of the collection database model
+ * @param {Object} [params.pdrPgModel] - Instance of the pdr database model
+ * @param {Object} [params.providerPgModel] - Instance of the provider database model
+ * @param {Object} [params.filePgModel] - Instance of the file database model
  * @returns {Object} An API Granule with associated Files
  */
 export const translatePostgresGranuleToApiGranule = async ({