+++ packages/db/tests/lib/test-granule.js
@@ -82,13 +82,14 @@ test('upsertGranuleWithExecutionJoinRecord() creates granule record with granule
     status: 'running',
   });
 
-  const [granuleCumulusId] = await knex.transaction(
+  const [pgGranule] = await knex.transaction(
     (trx) => upsertGranuleWithExecutionJoinRecord(
       trx,
       granule,
       executionCumulusId
     )
   );
+  const granuleCumulusId = pgGranule.cumulus_id;
 
   const granuleRecord = await granulePgModel.get(
     knex,
@@ -129,13 +130,14 @@ test('upsertGranuleWithExecutionJoinRecord() handles multiple executions for a g
     status: 'completed',
   });
 
-  const [granuleCumulusId] = await knex.transaction(
+  const [pgGranule] = await knex.transaction(
     (trx) => upsertGranuleWithExecutionJoinRecord(
       trx,
       granule,
       executionCumulusId
     )
   );
+  const granuleCumulusId = pgGranule.cumulus_id;
 
   const [secondExecutionCumulus] = await executionPgModel.create(
     knex,
@@ -240,11 +242,12 @@ test('upsertGranuleWithExecutionJoinRecord() will allow a running status to repl
     status: 'completed',
   });
 
-  const [granuleCumulusId] = await upsertGranuleWithExecutionJoinRecord(
+  const [pgGranule] = await upsertGranuleWithExecutionJoinRecord(
     knex,
     granule,
     executionCumulusId
   );
+  const granuleCumulusId = pgGranule.cumulus_id;
 
   const [secondExecution] = await executionPgModel.create(
     knex,
@@ -301,11 +304,12 @@ test('upsertGranuleWithExecutionJoinRecord() succeeds if granulePgModel.upsert()
     status: 'completed',
   });
 
-  const [granuleCumulusId] = await upsertGranuleWithExecutionJoinRecord(
+  const [pgGranule] = await upsertGranuleWithExecutionJoinRecord(
     knex,
     granule,
     executionCumulusId
   );
+  const granuleCumulusId = pgGranule.cumulus_id;
 
   const updatedGranule = {
     ...granule,
@@ -945,9 +949,9 @@ test('getUniqueGranuleByGranuleId() returns a single granule', async (t) => {
   const fakeGranule = fakeGranuleRecordFactory({
     collection_cumulus_id: collectionCumulusId,
   });
-  const [granuleCumulusId] = await granulePgModel.create(knex, fakeGranule);
+  const [createdPgGranule] = await granulePgModel.create(knex, fakeGranule);
 
-  const pgGranule = await granulePgModel.get(knex, { cumulus_id: granuleCumulusId });
+  const pgGranule = await granulePgModel.get(knex, { cumulus_id: createdPgGranule.cumulus_id });
 
   t.deepEqual(
     await getUniqueGranuleByGranuleId(knex, pgGranule.granule_id, granulePgModel),
@@ -981,10 +985,11 @@ test('getUniqueGranuleByGranuleId() throws an error if more than one granule is 
     }),
   ];
 
-  const granuleIds = await Promise.all(fakeGranules.map((fakeGranule) =>
+  const granules = await Promise.all(fakeGranules.map((fakeGranule) =>
     granulePgModel.create(knex, fakeGranule)));
+  const granuleIds = granules.map(([granule]) => granule.cumulus_id);
 
-  const pgGranule = await granulePgModel.get(knex, { cumulus_id: granuleIds[0][0] });
+  const pgGranule = await granulePgModel.get(knex, { cumulus_id: granuleIds[0] });
 
   await t.throwsAsync(
     getUniqueGranuleByGranuleId(knex, pgGranule.granule_id, granulePgModel),