+++ tf-modules/archive/ingest-reporting.tf
@@ -219,13 +219,6 @@ resource "aws_sns_topic" "report_granules_topic" {
   tags = var.tags
 }
 
-resource "aws_lambda_event_source_mapping" "publish_granules" {
-  event_source_arn  = data.aws_dynamodb_table.granules.stream_arn
-  function_name     = aws_lambda_function.publish_granules.arn
-  starting_position = "TRIM_HORIZON"
-  batch_size        = 10
-}
-
 # Report PDRs
 
 resource "aws_iam_role" "publish_pdrs_lambda_role" {